⚡️ ファイル名の変更とリファクタリング ファイル名: auth_page -> base_page
